Додано зміну формату з 24-годинного у 12-годинний
